+++ src/Main/Bind/BrowserViewBind.ts
@@ -1,9 +1,10 @@
-import {BrowserWindow, BrowserView, shell, clipboard, Menu, MenuItem, Rectangle} from 'electron';
+import {BrowserView, BrowserWindow, clipboard, Menu, MenuItem, Rectangle, shell} from 'electron';
 import fs from 'fs';
 import os from 'os';
-import path from "path";
+import path from 'path';
 import {ShellUtil} from '../../Renderer/Library/Util/ShellUtil';
 import {BrowserViewIPC} from '../../IPC/BrowserViewIPC';
+import BrowserWindowConstructorOptions = Electron.BrowserWindowConstructorOptions;
 
 class _BrowserViewBind {
   private window: BrowserWindow;
@@ -30,7 +31,7 @@ class _BrowserViewBind {
     });
 
     // bind IPC
-    BrowserViewIPC.initWindow(window)
+    BrowserViewIPC.initWindow(window);
     BrowserViewIPC.onLoadURL(async (_ev, url) => this.loadURL(url));
     BrowserViewIPC.onGetURL(() => this.getURL());
     BrowserViewIPC.onHide((_ev, flag) => this.hide(flag));
@@ -42,13 +43,15 @@ class _BrowserViewBind {
     BrowserViewIPC.onFocus(async () => webContents.focus());
     BrowserViewIPC.onBlur(async () => window.webContents.focus());
     BrowserViewIPC.onExecuteJavaScript((_ev, js) => webContents.executeJavaScript(js));
-    BrowserViewIPC.onInsertCSS((_ev, css) => { webContents.insertCSS(css); }); // 値を返却するとエラーになるので{}で囲む
+    BrowserViewIPC.onInsertCSS((_ev, css) => {
+      webContents.insertCSS(css);
+    }); // 値を返却するとエラーになるので{}で囲む
     BrowserViewIPC.onFindInPage((_ev, keyword, options) => {
       if (keyword) return webContents.findInPage(keyword, options);
     });
     BrowserViewIPC.onStopFindInPage((_ev, action) => webContents.stopFindInPage(action));
-    BrowserViewIPC.onSetRect((x, y, width, height) => this.setRect(x, y, width, height))
-    BrowserViewIPC.onSetBackgroundColor(color => this.browserView.setBackgroundColor(color))
+    BrowserViewIPC.onSetRect((x, y, width, height) => this.setRect(x, y, width, height));
+    BrowserViewIPC.onSetBackgroundColor(color => this.browserView.setBackgroundColor(color));
 
     const webContents = this.browserView.webContents;
     webContents.addListener('console-message', (_ev, level, message) => BrowserViewIPC.eventConsoleMessage(level, message));
@@ -59,6 +62,8 @@ class _BrowserViewBind {
     webContents.addListener('before-input-event', (_ev, input) => BrowserViewIPC.eventBeforeInput(input));
     webContents.addListener('found-in-page', (_ev, result) => BrowserViewIPC.eventFoundInPage(result));
     webContents.session.on('will-download', () => BrowserViewIPC.eventWillDownload());
+    // github projectでissueをクリックしたときにmodal windowで開くようにするため。
+    webContents.setWindowOpenHandler(() => this.handleOpenNewWindow());
   }
 
   private setupContextMenu() {
@@ -91,7 +96,7 @@ class _BrowserViewBind {
         menu.append(new MenuItem({label: 'Cut text', click: () => webContents.cut()}));
       }
 
-      menu.append(new MenuItem({label: 'Paste text', click: ()=> webContents.paste()}));
+      menu.append(new MenuItem({label: 'Paste text', click: () => webContents.paste()}));
       menu.popup({window: this.window});
     });
   }
@@ -154,6 +159,17 @@ class _BrowserViewBind {
         this.window.setBrowserView(this.browserView);
       }
     }
+  }
+
+  private handleOpenNewWindow(): { action: 'allow'; overrideBrowserWindowOptions: BrowserWindowConstructorOptions } {
+    const [width, height] = this.window.getSize();
+    const overrideBrowserWindowOptions: BrowserWindowConstructorOptions = {
+      width: width - 100,
+      height: height - 100,
+      parent: this.window,
+    };
+
+    return {action: 'allow', overrideBrowserWindowOptions};
   }
 }