+++ src/Renderer/Fragment/Stream/UserStream/ProjectStreamEditorFragment.tsx
@@ -21,6 +21,7 @@ import {SampleIconNames} from '../SampleIconNames';
 import {Link} from '../../../Library/View/Link';
 import {ShellUtil} from '../../../Library/Util/ShellUtil';
 import {DocsUtil} from '../../../Library/Util/DocsUtil';
+import {GitHubV4ProjectNextClient} from '../../../Library/GitHub/V4/GitHubV4ProjectNextClient';
 
 type Props = {
   show: boolean;
@@ -53,7 +54,7 @@ export class ProjectStreamEditorFragment extends React.Component<Props, State> {
     errorProjectUrl: false,
     errorColor: false,
     errorIconName: false,
-  }
+  };
 
   componentDidUpdate(prevProps: Readonly<Props>, _prevState: Readonly<State>, _snapshot?: any) {
     // 表示されたときに初期化する
@@ -89,9 +90,45 @@ export class ProjectStreamEditorFragment extends React.Component<Props, State> {
     }
   }
 
+  // projectに関連するfilter streamを自動的に作成する。
+  // - iterationフィールドに基づいたフィルター
+  // - statusフィールドに基づいたフィルター
+  private async createFilterStream(projectStream: StreamEntity) {
+    // create client
+    const github = UserPrefRepo.getPref().github;
+    const gheVersion = UserPrefRepo.getGHEVersion();
+    const client = new GitHubV4ProjectNextClient(github.accessToken, github.host, github.https, gheVersion);
+
+    // get iterationName and statusName
+    const projectUrl = projectStream.queries[0];
+    const {error, iterationName, statusNames} = await client.getProjectStatusFieldNames(projectUrl);
+    if (error != null) {
+      console.error(error);
+      return;
+    }
+
+    // create iteration filter
+    {
+      const {error} = await StreamRepo.createStream('FilterStream', projectStream.id, `Current ${iterationName}`, [], `project-field:"${iterationName}/@current_iteration"`, projectStream.notification, projectStream.color);
+      if (error != null) {
+        console.error(error);
+        return;
+      }
+    }
+
+    // create status filter
+    for (const statusName of statusNames) {
+      const {error} = await StreamRepo.createStream('FilterStream', projectStream.id, statusName, [], `project-field:"status/${statusName}"`, projectStream.notification, projectStream.color);
+      if (error != null) {
+        console.error(error);
+        return;
+      }
+    }
+  }
+
   private async handleEdit() {
     const name = this.state.name?.trim();
-    const projectUrl = this.state.projectUrl.trim();
+    const projectUrl = this.state.projectUrl.trim().replace(/\/views\/\d+$/, ''); // beta projectの場合、URL末尾に/view/1のようにつくことがある。正規化のためにこれを削除しておく。
     const color = this.state.color?.trim();
     const notification = this.state.notification ? 1 : 0;
     const iconName = this.state.iconName?.trim() as IconNameType;
@@ -112,6 +149,7 @@ export class ProjectStreamEditorFragment extends React.Component<Props, State> {
     } else {
       const {error, stream} = await StreamRepo.createStream('ProjectStream', null, name, [projectUrl], '', notification, color, iconName);
       if (error) return console.error(error);
+      await this.createFilterStream(stream);
       this.props.onClose(true, stream.id);
     }
   }