+++ src/Renderer/Library/View/IssueRow.tsx
@@ -21,6 +21,7 @@ import {IconButton} from './IconButton';
 import {PlatformUtil} from '../Util/PlatformUtil';
 import {UserIcon} from './UserIcon';
 import {ShellUtil} from '../Util/ShellUtil';
+import {RemoteProjectFieldEntity} from '../Type/RemoteGitHubV3/RemoteIssueEntity';
 
 type Props = {
   issue: IssueEntity;
@@ -39,6 +40,7 @@ type Props = {
   onToggleRead?: (issue: IssueEntity) => void;
   onToggleIssueType?: (issue: IssueEntity) => void;
   onToggleProject?: (issue: IssueEntity, projectName: string, projectColumn: string) => void;
+  onToggleProjectField?: (issue: IssueEntity, projectField: RemoteProjectFieldEntity) => void;
   onToggleMilestone?: (issue: IssueEntity) => void;
   onToggleLabel?: (issue: IssueEntity, label: string) => void;
   onToggleAuthor?: (issue: IssueEntity) => void;
@@ -59,10 +61,10 @@ type State = {
 export class IssueRow extends React.Component<Props, State> {
   state: State = {
     showMenu: false,
-  }
+  };
 
   private contextMenus: ContextMenuType[] = [];
-  private contextMenuPos: {left: number; top: number};
+  private contextMenuPos: { left: number; top: number };
   private contextMenuHorizontalLeft: boolean;
   private contextMenuHideBrowserView: boolean;
 
@@ -95,6 +97,7 @@ export class IssueRow extends React.Component<Props, State> {
       if (nextProps.onToggleRead !== this.props.onToggleRead) return true;
       if (nextProps.onToggleIssueType !== this.props.onToggleIssueType) return true;
       if (nextProps.onToggleProject !== this.props.onToggleProject) return true;
+      if (nextProps.onToggleProjectField !== this.props.onToggleProjectField) return true;
       if (nextProps.onToggleMilestone !== this.props.onToggleMilestone) return true;
       if (nextProps.onToggleLabel !== this.props.onToggleLabel) return true;
       if (nextProps.onToggleAuthor !== this.props.onToggleAuthor) return true;
@@ -120,7 +123,7 @@ export class IssueRow extends React.Component<Props, State> {
   }
 
   private isOpenRequest(ev: React.MouseEvent): boolean {
-    return !!(ev.shiftKey || ev.metaKey)
+    return !!(ev.shiftKey || ev.metaKey);
   }
 
   private isFilterToggleRequest(ev: React.MouseEvent): boolean {
@@ -242,7 +245,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Issue State', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleIssueType(this.props.issue)},
+      {label: 'Filter Issue State', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleIssueType(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Issues', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openIssues()},
     ];
@@ -267,9 +270,39 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Project', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleProject(this.props.issue, projectName, projectColumn)},
+      {
+        label: 'Filter Project',
+        subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,
+        icon: 'filter-outline',
+        handler: () => this.props.onToggleProject(this.props.issue, projectName, projectColumn)
+      },
       {type: 'separator'},
       {label: 'Open Project', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openProject(projectUrl)},
+    ];
+
+    this.contextMenuHideBrowserView = true;
+    this.contextMenuHorizontalLeft = false;
+    this.contextMenuPos = {top: ev.clientY, left: ev.clientX};
+    this.setState({showMenu: true});
+  }
+
+  private handleContextMenuProjectField(ev: React.MouseEvent, projectField: RemoteProjectFieldEntity) {
+    if (this.props.disableMenu) return;
+
+    if (this.isFilterToggleRequest(ev)) {
+      this.props.onToggleProjectField(this.props.issue, projectField);
+      return;
+    }
+
+    if (this.isOpenRequest(ev)) {
+      this.openProject(projectField.projectUrl);
+      return;
+    }
+
+    this.contextMenus = [
+      {label: 'Filter Project Field', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleProjectField(this.props.issue, projectField)},
+      {type: 'separator'},
+      {label: 'Open Project', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openProject(projectField.projectUrl)},
     ];
 
     this.contextMenuHideBrowserView = true;
@@ -292,7 +325,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Milestone', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleMilestone(this.props.issue)},
+      {label: 'Filter Milestone', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleMilestone(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Milestone', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openMilestone()},
     ];
@@ -317,7 +350,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Label', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleLabel(this.props.issue, label)},
+      {label: 'Filter Label', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleLabel(this.props.issue, label)},
       {type: 'separator'},
       {label: 'Open Label', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openLabel(label)},
     ];
@@ -342,7 +375,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Author', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleAuthor(this.props.issue)},
+      {label: 'Filter Author', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleAuthor(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Author', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openUser(this.props.issue.author)},
     ];
@@ -367,7 +400,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Assignee', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleAssignee(this.props.issue, loginName)},
+      {label: 'Filter Assignee', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleAssignee(this.props.issue, loginName)},
       {type: 'separator'},
       {label: 'Open Assignee', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openUser(loginName)},
     ];
@@ -392,7 +425,12 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Review Requested', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleReviewRequested(this.props.issue, loginName)},
+      {
+        label: 'Filter Review Requested',
+        subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,
+        icon: 'filter-outline',
+        handler: () => this.props.onToggleReviewRequested(this.props.issue, loginName)
+      },
       {type: 'separator'},
       {label: 'Open Review Requested', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openUser(loginName)},
     ];
@@ -417,7 +455,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Review', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleReview(this.props.issue, loginName)},
+      {label: 'Filter Review', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleReview(this.props.issue, loginName)},
       {type: 'separator'},
       {label: 'Open Review', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openUser(loginName)},
     ];
@@ -442,7 +480,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Org/User', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleRepoOrg(this.props.issue)},
+      {label: 'Filter Org/User', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleRepoOrg(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Org/User', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openOrg()},
     ];
@@ -467,7 +505,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Repository', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleRepoName(this.props.issue)},
+      {label: 'Filter Repository', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleRepoName(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Repository', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openRepo()},
     ];
@@ -492,7 +530,7 @@ export class IssueRow extends React.Component<Props, State> {
     }
 
     this.contextMenus = [
-      {label: 'Filter Number', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`,  icon: 'filter-outline', handler: () => this.props.onToggleIssueNumber(this.props.issue)},
+      {label: 'Filter Number', subLabel: `(${PlatformUtil.select('⌥', 'Alt')} Click)`, icon: 'filter-outline', handler: () => this.props.onToggleIssueNumber(this.props.issue)},
       {type: 'separator'},
       {label: 'Open Issue/PR', subLabel: `(${PlatformUtil.select('⌘', 'Shift')} Click)`, icon: 'open-in-new', handler: () => this.openIssue()},
     ];
@@ -556,7 +594,7 @@ export class IssueRow extends React.Component<Props, State> {
     const readClassName = IssueRepo.isRead(this.props.issue) ? 'issue-read' : 'issue-unread';
     const selectedClassName = this.props.selected ? 'issue-selected' : 'issue-unselected';
     const fadeInClassName = this.props.fadeIn ? 'issue-fadein' : '';
-    const slimClassName = this.props.slim ? 'issue-slim' : ''
+    const slimClassName = this.props.slim ? 'issue-slim' : '';
 
     return (
       <Root
@@ -593,7 +631,7 @@ export class IssueRow extends React.Component<Props, State> {
 
     let warningMergeable;
     if (issue.value.mergeable === 'CONFLICTING') {
-      warningMergeable = <WarningMergeableIcon name='exclamation-thick' color={color.white} size={iconFont.nano}/>
+      warningMergeable = <WarningMergeableIcon name='exclamation-thick' color={color.white} size={iconFont.nano}/>;
     }
 
     return (
@@ -619,6 +657,7 @@ export class IssueRow extends React.Component<Props, State> {
     return (
       <Attributes>
         {this.renderProjects()}
+        {this.renderProjectFields()}
         {this.renderMilestone()}
         {this.renderLabels()}
       </Attributes>
@@ -630,7 +669,7 @@ export class IssueRow extends React.Component<Props, State> {
     if (!projects?.length) return;
 
     const projectViews = projects.map((project, index) => {
-      const label = `${project.name}:${project.column}`
+      const label = `${project.name}:${project.column}`;
       return (
         <Project
           // onClick={(ev) => this.handleClickProject(ev, project.name, project.column, project.url)}
@@ -648,6 +687,47 @@ export class IssueRow extends React.Component<Props, State> {
     return (
       <React.Fragment>
         {projectViews}
+      </React.Fragment>
+    );
+  }
+
+  private renderProjectFields() {
+    const projectFields = this.props.issue.value.projectFields;
+    if (!projectFields?.length) return;
+
+    const projectFieldViews = projectFields.map((projectField, index) => {
+      // プロジェクトに入っているissueはかならずtitleフィールドが含まれる。表示する必要はないのでスキップする。
+      if (projectField.name.toLocaleLowerCase() === 'title') return null;
+
+      // expanded_iterationはfilterのために展開されたiterationである。これは表示には不要なのでスキップする。
+      if (projectField.dataType === 'EXPANDED_ITERATION') return null;
+
+      let iconName = ({
+        SINGLE_SELECT: 'format-list-checkbox',
+        ITERATION: 'refresh',
+        TEXT: 'format-text',
+        // NUMBER: 'numeric',
+        NUMBER: 'counter',
+        DATE: 'calendar-blank-outline'
+      })[projectField.dataType] ?? 'border-none-variant';
+      if (projectField.name.toLocaleLowerCase() === 'status') iconName = 'view-column-outline';
+
+      return (
+        <Project
+          onClick={ev => this.handleContextMenuProjectField(ev, projectField)}
+          onContextMenu={ev => this.handleContextMenuProjectField(ev, projectField)}
+          title={`${projectField.projectTitle}/${projectField.name}/${projectField.value} (Ctrl + Click)`}
+          key={index}
+        >
+          <Icon name={iconName} size={iconFont.small}/>
+          <ProjectFieldText singleLine={true}>{projectField.value}</ProjectFieldText>
+        </Project>
+      );
+    });
+
+    return (
+      <React.Fragment>
+        {projectFieldViews}
       </React.Fragment>
     );
   }
@@ -697,7 +777,7 @@ export class IssueRow extends React.Component<Props, State> {
   }
 
   private renderUsers() {
-    const updated  = DateUtil.localToString(new Date(this.props.issue.value.updated_at));
+    const updated = DateUtil.localToString(new Date(this.props.issue.value.updated_at));
     const read = DateUtil.localToString(new Date(this.props.issue.read_at));
     const iconColor = this.props.selected ? color.white : appTheme().icon.soft;
 
@@ -738,7 +818,7 @@ export class IssueRow extends React.Component<Props, State> {
         >
           <Image source={{url: assignee.avatar_url}}/>
         </Assignee>
-      )
+      );
     });
 
     return (
@@ -753,7 +833,7 @@ export class IssueRow extends React.Component<Props, State> {
     const author = this.props.issue.author;
     const reviews = this.props.issue.value.reviews
       ?.filter(review => review.login !== author)
-      ?.sort((a, b) =>  b.state > a.state ? 1 : -1); //  'COMMENTED', 'CHANGES_REQUESTED', 'APPROVED' の順番に表示する
+      ?.sort((a, b) => b.state > a.state ? 1 : -1); //  'COMMENTED', 'CHANGES_REQUESTED', 'APPROVED' の順番に表示する
 
     const reviewRequested = this.props.issue.value.requested_reviewers;
 
@@ -818,7 +898,7 @@ export class IssueRow extends React.Component<Props, State> {
     const {repoOrg, repoName} = GitHubUtil.getInfo(this.props.issue.value.url);
 
     const date = new Date(this.props.issue.value.updated_at);
-    const updated  = DateUtil.localToString(date);
+    const updated = DateUtil.localToString(date);
     const read = DateUtil.localToString(new Date(this.props.issue.read_at));
 
     let privateIcon;
@@ -1050,6 +1130,12 @@ const ProjectText = styled(Text)`
   font-weight: ${fontWeight.softBold};
   padding-left: ${space.tiny}px;
   max-width: 100px;
+`;
+
+const ProjectFieldText = styled(Text)`
+  font-size: ${font.small}px;
+  font-weight: ${fontWeight.softBold};
+  padding-left: ${space.tiny}px;
 `;
 
 const Milestone = styled(ClickView)`